requirements.txt
beautifulsoup4
lxml
pandas
psycopg2-binary
pytz
requests
spotipy
sqlalchemy
streamlit
boto3
python-dotenv
